feat : resolve singletone design
