+++ src/comm/cache_index.cpp
@@ -38,26 +38,24 @@ int8_t CacheIndex::GetFreeIndex(const uint8_t livox_lidar_type, const uint32_t h
   if (ret != 0) {
     return -1;
   }
-  {
-    std::lock_guard<std::mutex> lock(index_mutex_);
-    if (map_index_.find(key) != map_index_.end()) {
-      index = map_index_[key];
+  
+  std::lock_guard<std::mutex> lock(index_mutex_);
+  // Check again inside the lock
+  if (map_index_.find(key) != map_index_.end()) {
+    index = map_index_[key];
+    return 0;
+  }
+
+  printf("GetFreeIndex key:%s.\n", key.c_str());
+  for (size_t i = 0; i < kMaxSourceLidar; ++i) {
+    if (!index_cache_[i]) {
+      index_cache_[i] = 1;
+      map_index_[key] = static_cast<uint8_t>(i);
+      index = static_cast<uint8_t>(i);
       return 0;
     }
   }
 
-  {
-    printf("GetFreeIndex key:%s.\n", key.c_str());
-    std::lock_guard<std::mutex> lock(index_mutex_);
-    for (size_t i = 0; i < kMaxSourceLidar; ++i) {
-      if (!index_cache_[i]) {
-        index_cache_[i] = 1;
-        map_index_[key] = static_cast<uint8_t>(i);
-        index = static_cast<uint8_t>(i);
-        return 0;
-      }
-    }
-  }
   return -1;
 }
 
@@ -78,11 +76,12 @@ int8_t CacheIndex::GetIndex(const uint8_t livox_lidar_type, const uint32_t handl
     return -1;
   }
 
+  std::lock_guard<std::mutex> lock(index_mutex_);
   if (map_index_.find(key) != map_index_.end()) {
-    std::lock_guard<std::mutex> lock(index_mutex_);
     index = map_index_[key];
     return 0;
   }
+  
   printf("Can not get index, the livox lidar type:%u, handle:%u\n", livox_lidar_type, handle);
   return -1;
 }
@@ -93,7 +92,8 @@ int8_t CacheIndex::LvxGetIndex(const uint8_t livox_lidar_type, const uint32_t ha
   if (ret != 0) {
     return -1;
   }
-
+  
+  // No lock here intentionally, as Lvx file processing is single-threaded
   if (map_index_.find(key) != map_index_.end()) {
     index = map_index_[key];
     return 0;
@@ -110,12 +110,14 @@ void CacheIndex::ResetIndex(LidarDevice *lidar) {
     return;
   }
 
+  std::lock_guard<std::mutex> lock(index_mutex_);
   if (map_index_.find(key) != map_index_.end()) {
     uint8_t index = map_index_[key];
-    std::lock_guard<std::mutex> lock(index_mutex_);
     map_index_.erase(key);
-    index_cache_[index] = 0;
+    if(index < kMaxSourceLidar) {
+      index_cache_[index] = 0;
+    }
   }
 }
 
-} // namespace
+} // namespace livox_ros
